Move fill impl into kernels::fill
