+++ candle-metal-kernels/src/kernels/mod.rs
@@ -2,6 +2,7 @@ pub mod affine;
 pub mod binary;
 pub mod cast;
 pub mod convolution;
+pub mod fill;
 pub mod indexing;
 mod macros;
 pub mod mlx_gemm;
@@ -17,6 +18,7 @@ pub use affine::*;
 pub use binary::{call_binary_contiguous, call_binary_strided};
 pub use cast::{call_cast_contiguous, call_cast_strided};
 pub use convolution::*;
+pub use fill::*;
 pub use indexing::*;
 pub use mlx_gemm::{call_mlx_gemm, GemmDType};
 pub use quantized::{call_quantized_matmul_mm_t, call_quantized_matmul_mv_t, GgmlDType};
